truth-seeker/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
trafilatura==1.9.0
python-dotenv==1.0.1
tldextract==5.1.2
duckduckgo-search==5.3.1
beautifulsoup4==4.12.3
lxml
pydantic==2.13.1
numpy
